+++ x/evmutil/keeper/msg_server_test.go
@@ -388,7 +388,6 @@ func (suite *MsgServerSuite) TestConvertCosmosCoinToERC20_InitialContractDeploy(
 
 			// verify success
 			suite.NoError(err)
-			suite.Commit()
 
 			initiator := sdk.MustAccAddressFromBech32(tc.msg.Initiator)
 			receiver := testutil.MustNewInternalEVMAddressFromString(tc.msg.Receiver)
@@ -417,6 +416,24 @@ func (suite *MsgServerSuite) TestConvertCosmosCoinToERC20_InitialContractDeploy(
 			erc20Balance, err := suite.Keeper.QueryERC20BalanceOf(suite.Ctx, contractAddress, receiver)
 			suite.NoError(err)
 			suite.Equal(tc.amountConverted.BigInt(), erc20Balance, "unexpected erc20 balance for receiver")
+
+			// msg server event
+			suite.EventsContains(suite.GetEvents(),
+				sdk.NewEvent(
+					sdk.EventTypeMessage,
+					sdk.NewAttribute(sdk.AttributeKeyModule, types.AttributeValueCategory),
+					sdk.NewAttribute(sdk.AttributeKeySender, initiator.String()),
+				))
+
+			// keeper event
+			suite.EventsContains(suite.GetEvents(),
+				sdk.NewEvent(
+					types.EventTypeConvertCosmosCoinToERC20,
+					sdk.NewAttribute(types.AttributeKeyInitiator, initiator.String()),
+					sdk.NewAttribute(types.AttributeKeyReceiver, receiver.String()),
+					sdk.NewAttribute(types.AttributeKeyERC20Address, contractAddress.Hex()),
+					sdk.NewAttribute(types.AttributeKeyAmount, tc.msg.Amount.String()),
+				))
 		})
 	}
 }